requirements.txt
google-genai==0.4.0
streamlit==1.41.1
ipython==8.31.0
requests==2.32.3